requirements.txt
bcrypt
certifi
cffi
charset-normalizer
click
cryptography
dnspython
Flask
Flask-Bcrypt
flask-swagger-ui
Flask-UUID
idna
itsdangerous
Jinja2
MarkupSafe
pycparser
PyJWT
pymongo
requests
urllib3
Werkzeug
